refactor: implement Dockerized chain initialization in e2e tests (#1330)
Closes: #XXX

## What is the purpose of the change

This is a follow-up PR to #1293 and builds upon its work. It is part of the e2e test chain upgrade epic #1235 

This PR introduces the ability to run chain initialization in a Docker container by running the following:
```
docker run -v < path >:/tmp/osmo-test osmosis-e2e-chain-init:debug --data-dir=/tmp/osmo-test
```
All chain data is placed at the given `< path >` that is mounted as a volume on the container. In addition, this PR introduces documentation about the current state of the e2e tests. 

## Brief change log

- [pull chain temp folder creation our of chain to e2e package](https://github.com/osmosis-labs/osmosis/pull/1330/commits/c175289ba8c09bc178eafcfcca36a053fade961e)
- [create image and makefile steps to initialize chain state](https://github.com/osmosis-labs/osmosis/pull/1330/commits/de89119bde788b03de15eaef855f40dacb38c780)
- [allow for running the upgrade initialization in Docker by providing a data dir](https://github.com/osmosis-labs/osmosis/pull/1330/commits/dabb683955b0645c61462089af89bf022072eca9)
- [improve abstractions for chain initialization and add README](https://github.com/osmosis-labs/osmosis/pull/1330/commits/cf0d8a33053638553a228e3467ad5b439b83bafe)
- improve README

## Testing and Verifying

- ran e2e tests locally a few times
- `make build-e2e-chain-init`
- `make docker-build-e2e-chain-init`
- `docker run -v /home/roman/cosmos/osmosis/tmp:/tmp/osmo-test osmosis-e2e-chain-init:debug --data-dir=/tmp/osmo-test`

All steps worked as desired

## Documentation and Release Note

  - Does this pull request introduce a new feature or user-facing behavior changes? no
  - Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`? no
  - How is the feature or change documented? improved `tests/e2e/README.md`

## Next Steps

The next step is to switch our chain initialization logic in `func (s *IntegrationTestSuite) configureChain(chainId string)` to use Dockertest and the newly introduced container. Then, mount the genesis and configs on the Osmosis containers, against which the e2e tests are executed