Add structs for Upgrades and Forks structs (#1379)
Closes: #1375 

## What is the purpose of the change

Simplifies a lot of the upgrades management logic / boilerplate. The goal is that every new upgrade should from the app.go level, just require a one line update. And all the complexity would otherwise be held to within each upgrades module.

This PR is marked as draft until Changelog update + docs update to the `app/upgrades` README.

## Brief change log

* Introduces an Upgrade and Fork struct
* Adapt existing upgrades to use it

## Testing and Verifying

This should be covered by the existing test for the v4 upgrade, which tests that upgrade handler execution happens.

## Documentation and Release Note

  - Does this pull request introduce a new feature or user-facing behavior changes? no
  - Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`? Needs to be Done
  - How is the feature or change documented? (not applicable   /   specification (`x/<module>/spec/`)  /  [Osmosis docs repo](https://github.com/osmosis-labs/docs)   /   not documented)