Add rounding behavior to our binary search logic, to ensure we always round correctly (#3252)
## What is the purpose of the change

Needed for stableswap correctness. We ensure that the binary search always rounds `k` in the correct direction.

I'm currently in the process of rewriting the tests, to make this directly testable, and then will do a spec update. However the code for the stableswap change I believe is sufficiently commented.

## Brief Changelog

- Add rounding behavior option to ErrTolerance compare
- Use this in binary searcher
- Add this to stableswap cfmmSolve
- Improve tests

## Testing and Verifying

Working on test improvements here.

# Documentation and Release Note

  - Does this pull request introduce a new feature or user-facing behavior changes? yes
  - Is a relevant changelog entry added to the `Unreleased` section in `CHANGELOG.md`? no
  - How is the feature or change documented? Stableswap Spec